refactor: GameEntity, Operator, Enemy

+++ arclights/src/main/java/com/arclights/models/Sniper.java
@@ -3,9 +3,9 @@ package com.arclights.models;
 public class Sniper extends Operator {
     
     // Grid coordinates instead of pixel coordinates
-    public int gridX, gridY;
+    public double gridX, gridY;
 
-    public Sniper(int gridX, int gridY) {
+    public Sniper(double gridX, double gridY) {
         // HP=500, ATK=25, Range=3 tiles, Cooldown=60 frames (~1 attack per second)
         super(gridX, gridY, 500, 25, 3, 60);
         this.gridX = gridX;
